Remove admin view for old EmailAddress model; integrate its functionality into admin view for new one
